requirements-query.txt
matplotlib
pymongo
pandas
requests
pyYAML